Refactored some selectors to remove circular dependencies
